refactor: rename `pb_dynamic_compiler` to `dynamic_compiler`
