+++ lib/grpc_mock/dynamic_compiler/eex_codegen.ex
@@ -1,4 +1,4 @@
-defmodule GrpcMock.EExCodeGen do
+defmodule GrpcMock.DynamicCompiler.EExCodeGen do
   use GrpcMock.Codegen
   require Logger
   alias GrpcMock.Codegen
