--- lib/grpc_mock/pb_dynamic_compiler/compile_status.ex
@@ -1,8 +1,0 @@
-defmodule GrpcMock.PbDynamicCompiler.CompileStatus do
-  defstruct [:status, :error]
-  @type status :: :failed | :finished
-  @type t :: %__MODULE__{
-          status: status(),
-          error: any() | nil
-        }
-end
